Added some project structure
